Category parser.
A parser for emoji-test.txt
Introducing the category object

Objects are now clustered by gender and activity. We can chose a given
type of emoji and then randomly select skin color. That way, emojis
with a lot of variants have equal weight to those with fewer, or say,
objects, which only have one variant. The result will be more variety
on the screen once implemented.