Getting rid of flags again.
The industry has just not figured this out yet for reasons that are
beyond my comprehension. Plus, when we’re short on flags, the
re-running of generateReducedListOfSVGs in loadTheGrid, I’m not 100%
sure that’s good programming, so better to just never invoke it, even
though it’s always worked until now. But who knows why a hack works,
you know?